AnimationComponent created. Basic completion for the test shape class.
The AnimationComponent was created and filled with the necessary information to work. Also the class where the functions of the ShapeComponent are being tested was modified by adding some basic data about the sprite node and their components. It's worth to mention that previously, this node and their components had "rectangle" in their names instead of "sprite".