Struct added to ShapeComponent class. AnimationComponent class finished. Use of color buffer in the class of shape test.
Struct added to the ShapeComponent class to group all the data that contains the texture, so when passing values to the animation class it doesnt need a lot of parameters.
The AnimationComponent class is now completed and does the same that the Animation class but addapted to the scene graph system.
The class created for testing all the functions that have to do with the ShapeComponent and AnimationComponent, now also has a node with a ShapeComponent to test all that is realted with the colo buffer.