Change how the camera and the scen grap systems work
The camera now takes in account the forward of it so when the camera moves, it goes in the direction that the forward points to.
The scene graph system -talking specifically about the load of models- now loads the mesh components and add childs in the correct way, making the correct hierarchy.